Updates thru NEH 13:1
Moved "named on the sealed document" into the notes instead of the attributes.
Corrected the chapter and verse reference for the "named on the sealed document" text.
Changed "THE" (definite article) Shilonite from Asaiah_4 to the prophet Ahijah_1.
Added notes regarding Levitical gatekeepers (for reference).
Added Bilgah_2 (NEH 12:5)
Added Obadiah_11 (NEH 12:25)
Added Joiakim_2 (based on lineage in NEH 12:26, it was a different Joiakim).