Updates thru ISA 13:1
Noted Urijah_1 as Levite.
Added King of Assyria_1 (existed in PersonLabel but not here).
Added Tabeel_2 (not the same person as Tabeel_1 who lived a few hundred years later during the time of Ezra).
Noted Immanuel as Isaiah's son who is the same person as Maher-shalal-hash-baz (so I removed him as a separate individual).